+++ prototyper/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. See [conven
 ## Unreleased
 
 ### Added
+- Add AIA IMSIC IPI backend support for RustSBI Prototyper.
 
 ### Modified
 - Refine CSR group comments.